+++ core-ui/src/main/java/com/agileburo/anytype/core_ui/common/Checkable.kt
@@ -4,5 +4,5 @@ package com.agileburo.anytype.core_ui.common
  * Defines a view that can be checked
  */
 interface Checkable {
-    val isChecked: Boolean
+    var isChecked: Boolean
 }