+++ domain/src/main/java/com/agileburo/anytype/domain/block/interactor/UpdateCheckbox.kt
@@ -1,26 +1,25 @@
 package com.agileburo.anytype.domain.block.interactor
 
 import com.agileburo.anytype.domain.base.BaseUseCase
-import com.agileburo.anytype.domain.base.Either
 import com.agileburo.anytype.domain.block.model.Command
 import com.agileburo.anytype.domain.block.repo.BlockRepository
+import com.agileburo.anytype.domain.event.model.Payload
 
+/**
+ * Use-case for updating checkbox state.
+ */
 open class UpdateCheckbox(
     private val repo: BlockRepository
-) : BaseUseCase<Unit, UpdateCheckbox.Params>() {
+) : BaseUseCase<Payload, UpdateCheckbox.Params>() {
 
-    override suspend fun run(params: Params) = try {
+    override suspend fun run(params: Params) = safe {
         repo.updateCheckbox(
             command = Command.UpdateCheckbox(
                 context = params.context,
                 target = params.target,
                 isChecked = params.isChecked
             )
-        ).let {
-            Either.Right(it)
-        }
-    } catch (t: Throwable) {
-        Either.Left(t)
+        )
     }
 
     /**
@@ -33,5 +32,4 @@ open class UpdateCheckbox(
         val target: String,
         val isChecked: Boolean
     )
-
 }