+++ presentation/src/main/java/com/agileburo/anytype/presentation/page/editor/Orchestrator.kt
@@ -212,7 +212,7 @@ class Orchestrator(
                         )
                     ).proceed(
                         failure = defaultOnError,
-                        success = {}
+                        success = defaultPayload
                     )
                 }
                 is Intent.Text.UpdateText -> {