+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Fixes & tech 🚒
 
+* Checkbox state in action-menu preview is not synced with editor state (#748)
+* User interactions with checkbox button are not synced correctly with the middleware (#749, #668)
 * When searching for pages, if filter text is empty space, query returns only pages where title contains empty spaces (#746)
 * Regression. Text is not always set when creating a lot of text blocks (#741)
 * Respective theme colors should differ for text color and background colors in action menu (#738)